Rename OAuth,DOCS and related files
