+++ app.py
@@ -45,7 +45,7 @@ from usersutils import valid_username
 
 sys.path.append(os.path.dirname(os.path.abspath(__file__)))
 
-from OAuth.config import ALLOWED_EMAILS, GOOGLE_CLIENT_ID
+from oauth.config import ALLOWED_EMAILS, GOOGLE_CLIENT_ID
 
 ALLOWED_EXTENSIONS = {'jpg', 'jpeg', 'png', 'gif', 'webp', 'heif', 'pdf'}
