+++ docs/authentication.md
