+++ docs/setup.md
